# contracts/openapi/public-api.yaml
openapi: 3.1.0
...
info:
  title: EduTwin Public API
  version: 2.0.0
  license:
    name: Course demonstration use
    identifier: LicenseRef-Course-Demonstration
  description: |
    EduTwin browser-facing contract. MySQL is the system of record. Redis values,
    task state, and server-sent events are recoverable projections. Every traced
    response identifies the accepted answer event, analysis job, preallocated or
    persisted snapshot, data versions, and requested and effective model versions.
    Authentication uses a stateless JWT with one active role and all granted roles.
    Role, password, and account changes advance the token version and invalidate older
    tokens. Logout consists of discarding the JWT in the browser.
servers:
  - url: http://127.0.0.1:8080
    description: Local Compose deployment
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      summary: Return the authenticated principal
      responses:
        '200':
          description: Current principal and authorized courses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/password:
    post:
      tags: [Authentication]
      operationId: changePassword
      summary: Change the authenticated account password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '204': { description: Password changed; the current token must be discarded. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/courses/{courseId}/practice/next:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Practice]
      operationId: getNextPracticeQuestion
      summary: Return the next rule-selected practice question
      description: Access is restricted to the authenticated student's enrollment.
      x-rbac: [STUDENT_SELF]
      responses:
        '200':
          description: Next question and the twin trace used for selection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PracticeQuestion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/courses/{courseId}/answers:
    parameters:
      - $ref: '#/components/parameters/CourseId'
    get:
      tags: [Practice]
      operationId: listAnswerHistory
      summary: Return the authenticated student's course answer history
      description: Results are ordered by eventSequence descending and include online and imported answer events.
      x-rbac: [STUDENT_SELF]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Paged answer history for the authenticated student and course.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerHistoryPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Practice]
      operationId: submitAnswer
      summary: Accept an answer for asynchronous analysis
      description: |
        The transaction writes the answer event, a QUEUED analysis job with a
        preallocated snapshot identifier, and an outbox event. The response does
        not wait for FastAPI, Redis consumption, or the configured LLM. Replaying the same
        key and canonical request returns the original job. Reusing the key with a
        different request returns 409.
      x-rbac: [STUDENT_SELF]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerSubmission'
      responses:
        '202':
          description: Answer persisted and analysis queued.
          headers:
            Location:
              required: true
              description: Relative URL of the analysis job.
              schema:
                type: string
            Retry-After:
              required: true
              description: Suggested polling delay in seconds.
              schema:
                type: integer
                minimum: 1
            Idempotency-Replayed:
              required: true
              description: Indicates that the original response was replayed.
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /api/v1/courses/{courseId}/practice/tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/CourseId'
      - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Practice]
      operationId: getLearningPlanPracticeTask
      summary: Resolve a current plan task to its practice question
      x-rbac: [STUDENT]
      responses:
        '200': { description: Practice task., content: { application/json: { schema: { $ref: '#/components/schemas/PlanPracticeTask' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Stateless JWT containing one role plus userId and authorization-version claims; logout discards the token client-side.
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Key scoped to the authenticated user and answer endpoint.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Authenticated principal lacks the required role or scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found within the authorized scope.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Idempotency conflict, stale snapshot, or immutable-version conflict.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    UnprocessableEntity:
      description: Request is syntactically valid but violates a domain invariant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    AuthenticatedUser:
      type: object
      additionalProperties: false
      required: [userId, username, displayName, role, activeRole, availableRoles, mustChangePassword, accessibleCourseIds]
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
          minLength: 1
        displayName:
          type: string
          minLength: 1
        role:
          $ref: '#/components/schemas/UserRole'
        activeRole:
          $ref: '#/components/schemas/UserRole'
        availableRoles:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/UserRole' }
        mustChangePassword:
          type: boolean
        accessibleCourseIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
    UserRole:
      type: string
      enum: [ADMIN, TEACHER, STUDENT, COUNSELOR]
    PasswordChangeRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 12 }
...
    DataVersionRef:
      type: object
      additionalProperties: false
      required: [sourceId, version, manifestSha256]
      properties:
        sourceId:
          type: string
          enum: [ASSISTMENTS_2009_2010_SKILL_BUILDER_CORRECTED, OULAD, EDUTWIN_DEMO]
        version:
          type: string
          minLength: 1
        manifestSha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
    ModelVersionRef:
      type: object
      additionalProperties: false
      required: [purpose, family, modelName, modelVersion, artifactSha256, calibratorVersion]
      properties:
        purpose:
          type: string
          enum: [MASTERY, NEXT_CORRECT, RISK, EXPLANATION, DIAGNOSIS, PLAN_RULES]
        family:
          type: string
          enum: [IRT, BKT, DKT, AKT, LOGISTIC_REGRESSION, LIGHTGBM, CATBOOST, SHAP, DEEPSEEK, OPENAI_COMPATIBLE, RULE, TEMPLATE]
        modelName:
          type: string
          minLength: 1
        modelVersion:
          type: string
          minLength: 1
        artifactSha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        calibratorVersion:
          type: [string, 'null']
          minLength: 1
    TraceRef:
      type: object
      additionalProperties: false
      required:
        - correlationId
        - answerEventId
        - analysisJobId
        - snapshotId
        - dataVersions
        - requestedModelVersions
        - effectiveModelVersions
      properties:
        correlationId:
          type: string
          format: uuid
        answerEventId:
          type: string
          format: uuid
        analysisJobId:
          type: string
          format: uuid
        snapshotId:
          type: string
          format: uuid
          description: Preallocated for QUEUED jobs and persisted unchanged on completion.
        dataVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/DataVersionRef'
        requestedModelVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ModelVersionRef'
        effectiveModelVersions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ModelVersionRef'
...
    PracticeChoice:
      type: object
      additionalProperties: false
      required: [choiceId, label]
      properties:
        choiceId:
          type: string
          minLength: 1
        label:
          type: string
          minLength: 1
    PracticeQuestion:
      type: object
      additionalProperties: false
      required: [questionId, courseId, prompt, questionType, choices, skillIds]
      properties:
        trace:
          $ref: '#/components/schemas/TraceRef'
          x-edutwin-visibility: internal
        questionId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        prompt:
          type: string
          minLength: 1
        questionType:
          type: string
          enum: [SINGLE_CHOICE]
        choices:
          type: array
          minItems: 2
          items:
            $ref: '#/components/schemas/PracticeChoice'
        skillIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            format: uuid
    AnswerSubmission:
      type: object
      additionalProperties: false
      required: [questionId, selectedChoiceId, occurredAt]
      properties:
        questionId:
          type: string
          format: uuid
        selectedChoiceId:
          type: string
          minLength: 1
        occurredAt:
          type: string
          format: date-time
    AnswerHistorySkill:
      type: object
      additionalProperties: false
      required: [skillId, name]
      properties:
        skillId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
    AnswerHistoryItem:
      type: object
      additionalProperties: false
      required:
        - answerEventId
        - questionId
        - prompt
        - selectedChoice
        - correctChoice
        - correct
        - skills
        - attemptNumber
        - eventSequence
        - responseTimeMs
        - occurredAt
        - sourceType
      properties:
        answerEventId:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
        prompt:
          type: string
          minLength: 1
        selectedChoice:
          $ref: '#/components/schemas/PracticeChoice'
        correctChoice:
          $ref: '#/components/schemas/PracticeChoice'
        correct:
          type: boolean
        skills:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AnswerHistorySkill'
        attemptNumber:
          type: integer
          minimum: 1
          maximum: 65535
        eventSequence:
          type: integer
          format: int64
          minimum: 1
        responseTimeMs:
          type: integer
          format: int64
          minimum: 0
        occurredAt:
          type: string
          format: date-time
        sourceType:
          type: string
          enum: [ONLINE, IMPORTED]
    AnswerHistoryPage:
      type: object
      additionalProperties: false
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AnswerHistoryItem'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
    AnalysisJobStatus:
      type: string
      enum: [QUEUED, PROCESSING, COMPLETED, FAILED]
    AnalysisStage:
      type: string
      enum: [QUEUED, MODEL_INFERENCE, SNAPSHOT_PERSISTENCE, PLAN_GENERATION, EVENT_PUBLICATION, COMPLETED, FAILED]
    AnalysisResultLinks:
      type: object
      additionalProperties: false
      required: [twinUrl, learningPlanUrl, diagnosisUrl]
      properties:
        twinUrl:
          type: [string, 'null']
        learningPlanUrl:
          type: [string, 'null']
        diagnosisUrl:
          type: [string, 'null']
    AnalysisFailure:
      type: object
      additionalProperties: false
      required: [errorCode, message, retryable, failedAt]
      properties:
        errorCode:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        retryable:
          type: boolean
        failedAt:
          type: string
          format: date-time
    AnalysisJob:
      type: object
      additionalProperties: false
      required:
        - jobId
        - studentId
        - courseId
        - status
        - stage
        - lastEventSequence
        - submittedAt
        - startedAt
        - completedAt
        - failure
        - result
      properties:
        trace:
          $ref: '#/components/schemas/TraceRef'
          x-edutwin-visibility: internal
        jobId:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/AnalysisJobStatus'
        stage:
          $ref: '#/components/schemas/AnalysisStage'
        lastEventSequence:
          type: integer
          format: int64
          minimum: 0
          description: Highest persisted analysis_job_event sequence visible to SSE recovery.
        submittedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        degraded:
          type: boolean
          x-edutwin-visibility: internal
        degradationReasons:
          type: array
          uniqueItems: true
          x-edutwin-visibility: internal
          items:
            type: string
            enum:
              - KNOWLEDGE_MODEL_UNAVAILABLE
              - RISK_MODEL_UNAVAILABLE
              - MODEL_TIMEOUT
              - MODEL_CONTRACT_ERROR
              - DEEPSEEK_UNAVAILABLE
              - DEEPSEEK_TIMEOUT
              - LLM_UNAVAILABLE
              - LLM_TIMEOUT
              - REDIS_RETRY
        failure:
          oneOf:
            - $ref: '#/components/schemas/AnalysisFailure'
            - type: 'null'
        result:
          $ref: '#/components/schemas/AnalysisResultLinks'
...
    PlanPracticeTask:
      type: object
      additionalProperties: false
      required: [planTaskId, planId, courseId, questionId, questionTitle, prompt, answerType, optionsJson, skillId, skillName, targetCount, completedCount]
      properties:
        planTaskId: { type: string, format: uuid }
        planId: { type: string, format: uuid }
        courseId: { type: string, format: uuid }
        questionId: { type: string, format: uuid }
        questionTitle: { type: string }
        prompt: { type: string }
        answerType: { type: string }
        optionsJson: { type: [string, 'null'] }
        skillId: { type: string, format: uuid }
        skillName: { type: string }
        targetCount: { type: integer, minimum: 1 }
        completedCount: { type: integer, minimum: 0 }
...
    FieldViolation:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
    ProblemDetail:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, code, traceId, violations]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
        instance:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
        traceId:
          type: string
          minLength: 1
        violations:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'